+++ clients/all.d.ts
@@ -378,3 +378,4 @@ export import Chatbot = require('./chatbot');
 export import TimestreamInfluxDB = require('./timestreaminfluxdb');
 export import CodeConnections = require('./codeconnections');
 export import Deadline = require('./deadline');
+export import ControlCatalog = require('./controlcatalog');