+++ clients/controlcatalog.js
@@ -1,0 +1,18 @@
+require('../lib/node_loader');
+var AWS = require('../lib/core');
+var Service = AWS.Service;
+var apiLoader = AWS.apiLoader;
+
+apiLoader.services['controlcatalog'] = {};
+AWS.ControlCatalog = Service.defineService('controlcatalog', ['2018-05-10']);
+Object.defineProperty(apiLoader.services['controlcatalog'], '2018-05-10', {
+  get: function get() {
+    var model = require('../apis/controlcatalog-2018-05-10.min.json');
+    model.paginators = require('../apis/controlcatalog-2018-05-10.paginators.json').pagination;
+    return model;
+  },
+  enumerable: true,
+  configurable: true
+});
+
+module.exports = AWS.ControlCatalog;