+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog for AWS SDK for JavaScript
-<!--LATEST=2.1594.0-->
+<!--LATEST=2.1595.0-->
 <!--ENTRYINSERT-->
+
+## 2.1595.0
+* feature: ControlCatalog: This is the initial SDK release for AWS Control Catalog, a central catalog for AWS managed controls. This release includes 3 new APIs - ListDomains, ListObjectives, and ListCommonControls - that vend high-level data to categorize controls across the AWS platform.
+* feature: Mgn: Added USE_SOURCE as default option to LaunchConfigurationTemplate bootMode parameter.
+* feature: NetworkMonitor: Updated the allowed monitorName length for CloudWatch Network Monitor.
 
 ## 2.1594.0
 * feature: QuickSight: Adding IAMIdentityCenterInstanceArn parameter to CreateAccountSubscription